feat: Building out first two screens HomeScreen and AddNewDeckScreen
Had to implement a tab navigation view for these two pages. Struggled with the keyboard, dismissing it as well as with KeyboardAvoidingView. Began work on the README adding to it as I am building this time. It was cool being able to use those vector-icons that come with expo out of the box. It was also annoying that I couldn't figure out how to keep avoiding my top elements to crash into the space occupied by the time and battery bar of the iPhone