feat: Struggled but figured out how to use React Navigation to navigate between tabs and stacks
Also created some seed data to populate the list of decks on the home page. Implementing basic routing functionality proved to be a real pain and especially working with the navigation prop which sometimes went missing but I had to figure out how to get it by using the built-in React Navigation wrapper function. Got good experience in passing parameters into the navigation prop and accessing from the destination route