feat: Finally implemented and fixed state and data persistence issues
Was having huge problems managing the state in the app and tried to do it without Redux at first. But due to the differences in mobile and web around when componentDidMount triggers, I found it was so difficult to get and set data for mobile. Whereas for web you would have no such difficulties as each time a new screen re-renders that counts as a componentDidMount but in mobile once the initial tab and stack navigators mounted they don't necessarily remount again when you move back and forth within a stack. This caused me all sorts of pain as I couldn't get the app to automatically re-render to reflect a change in state in the appData. So I caved and decided to use Redux though if this was a web project I believe I could have easily got by without Redux as the data for this app isn't as complicated as in Project 2. Faced complexities setting up Redux for mobile especially in how to set up the Chrome Redux Extension Dev Tools (still can't get it to work) so I had to rely on a whole heap of console.log statements to debug. Finally I faced difficulties in getting both my app state in Redux and the app persisted data in AsyncStorage to be in sync due to the differences in handling data in each