feat: Implemented AddNewDeck feature flow
Was relatively straight forward as I could model it off the AddNewCard flow that I built previously. The only complication was when I went back to the AddNewDeckScreen it would always still keep the previous state of user input and not re-initialize and start blank again as the AddNewCardScreen would. Thought of some complicated ways to fix it but then realized that I could achieve the same outcome with something as simple as just handling the re-initialization upon click of the Add New Deck button