feat: Implemented Quiz feature flow and general style clean up
Was quite tricky and there were many considerations. Had a lot of trouble with styles but learnt a very handy way of debugging by chucking in backgroundColor props to View elements which were a lifesaver! Almost reaching the end now only need to fix styles in various places. Then review spec and compare what I have against it. Then test it all and submit! Thank God this was a very difficult project and I'm glad to be driving it home soon. I even had doubts at some point as to whether I could finish it, and even if I could, could I do it on time given I'm working full time in parallel