feat: Implemented daily local notifications feature via Expo Notifications and Permission API
initially I could not get the notifications to work. Which was very frustrating as it was the final feature I needed to meet spec and complete this project! I thought there was something wrong with my code but upon debugging I kept finding that when I provided the key to where I stored the notifications in AsyncStorage it kept returning null. Then I realized it was because of a permissions issue. The popup asking the user to grant permission never comes up for me in the simulation hence that branch of code I specified to set a value to the notifications key NEVER RUNS. Also did some styling fixes